roles/docker: force updating mervin configuration
It seems like Limnoria rewrites certain lists in its config in a
random order at runtime, so if it's done this then the configuration
won't be able to update. This unfortunately has the drawback that
mervin will *always* be restarted because the configuration *always*
updates, so the best place to fix this is probably in Limnoria itself.

Signed-off-by: John Helmert III <jchelmert3@posteo.net>