roles/services: use unique substrings for variable names
The contents of all variables with "system_services" in their name got
included in the list of services to enable, which also unfortunately
included variables that were intended to be stopped, because
"system_services" is a substring of "stop_system_services". This led
to services being started only to be stopped again. Use unique
substrings in variable names to avoid triggering this issue.

Signed-off-by: John Helmert III <jchelmert3@posteo.net>