roles/consul: don't leave on termination
By default, if a host is wrongly shut down, consul will gracefully
leave the cluster and we lose the node's consul and all of the
services with it, leaving no good way to alert for the event. With
this change, the node will not gracefully leave the cluster and it
will appear to be unhealthy to the other nodes (and their exporters).

Signed-off-by: John Helmert III <jchelmert3@posteo.net>